Add L1-Detail time information for single PLP.
System time is captured at initialization and will not be exactly
synchronized to real time when the flow starts. Subsequent time
information is calculated for each frame and may eventually not
be accurate due to floating point error accumulation. This can
be mitigated by using time-aligned frames or a frame time that's
naturally on a millisecond boundary.

Also note that although TAI time is attempted to be captured, many
Linux systems (especially those with systemd) will return UTC time.